model field

+++ bot/models.py
@@ -52,15 +52,15 @@ class User(models.Model):
     blocked = models.BooleanField(default=False, null=True)
 
     utm_source = models.CharField(max_length=50, null=True, blank=True)
-    utm_campaign = models.IntegerField(null=True, blank=True)
+    utm_campaign = models.BigIntegerField(null=True, blank=True)
 
-    phrase_id = models.IntegerField(null=True, blank=True)
+    phrase_id = models.BigIntegerField(null=True, blank=True)
     # utm_content
 
     device_type = models.CharField(max_length=50, null=True, blank=True)
 
 
-    ad_id = models.IntegerField(null=True, blank=True)
+    ad_id = models.BigIntegerField(null=True, blank=True)
 
 
     class Meta:
